feat(webhooks): remove deprecated System B webhook code path (v1.2.0)
Final removal of the System B webhook path introduced by
notification_dispatcher. v1.1.9 shipped the Webhooks v2 wire
contract; deprecation telemetry has been live since then; v1.2.0
drops the deprecated code entirely.

## Removed

- backend/src/api/handlers/notifications.rs (773 lines deleted) -
  the /api/v1/repositories/:id/notifications endpoints. Writes were
  already error'd-out in v1.1.9; this removes the routes entirely.
- backend/src/services/notification_dispatcher.rs (1008 lines deleted)
  including deliver_webhook, the webhook-channel branch in
  dispatch_event, and compute_webhook_signature (the v1.1.9 shared
  signing module replaces it).

## Added

- backend/migrations/086_drop_notification_subscriptions.sql - drops
  the deprecated table. v1.1.9 migrated data out; this removes the
  storage.
- backend/src/services/email_dispatcher.rs - the email side of the
  old notification dispatcher, moved into its own clean module so
  email functionality is no longer co-mingled with webhooks.

## Wiring

- main.rs unhooks notification_dispatcher; wires email_dispatcher
- openapi.rs drops the notifications module from the spec
- repositories.rs removes the legacy callsites
- webhook_producer.rs minor cleanup

## Tests

8981 backend lib tests pass. Clippy clean with -D warnings; cargo
fmt --check clean.

Closes #920.
